f/aws_arczonalshift_zonal_autoshift_configuration: Add rollback to create, remove null checks
```console
% make t T=TestAccARCZonalShiftZonalAutoshiftConfiguration_ K=arczonalshift
make: Verifying source code with gofmt...
==> Checking that code complies with gofmt requirements...
make: Running acceptance tests on branch: 🌿 f-aws_arczonalshift_zonal_autoshift_configuration 🌿...
TF_ACC=1 go1.26.2 test ./internal/service/arczonalshift/... -v -count 1 -parallel 20 -run='TestAccARCZonalShiftZonalAutoshiftConfiguration_'  -timeout 360m -vet=off
2026/04/23 15:53:06 Creating Terraform AWS Provider (SDKv2-style)...
2026/04/23 15:53:06 Initializing Terraform AWS Provider (SDKv2-style)...
=== RUN   TestAccARCZonalShiftZonalAutoshiftConfiguration_basic
=== PAUSE TestAccARCZonalShiftZonalAutoshiftConfiguration_basic
=== RUN   TestAccARCZonalShiftZonalAutoshiftConfiguration_disappears
=== PAUSE TestAccARCZonalShiftZonalAutoshiftConfiguration_disappears
=== RUN   TestAccARCZonalShiftZonalAutoshiftConfiguration_update
=== PAUSE TestAccARCZonalShiftZonalAutoshiftConfiguration_update
=== RUN   TestAccARCZonalShiftZonalAutoshiftConfiguration_blockingAlarms
=== PAUSE TestAccARCZonalShiftZonalAutoshiftConfiguration_blockingAlarms
=== RUN   TestAccARCZonalShiftZonalAutoshiftConfiguration_blockedWindows
=== PAUSE TestAccARCZonalShiftZonalAutoshiftConfiguration_blockedWindows
=== RUN   TestAccARCZonalShiftZonalAutoshiftConfiguration_allowedWindows
=== PAUSE TestAccARCZonalShiftZonalAutoshiftConfiguration_allowedWindows
=== RUN   TestAccARCZonalShiftZonalAutoshiftConfiguration_blockedDates
=== PAUSE TestAccARCZonalShiftZonalAutoshiftConfiguration_blockedDates
=== CONT  TestAccARCZonalShiftZonalAutoshiftConfiguration_basic
=== CONT  TestAccARCZonalShiftZonalAutoshiftConfiguration_blockedWindows
=== CONT  TestAccARCZonalShiftZonalAutoshiftConfiguration_update
=== CONT  TestAccARCZonalShiftZonalAutoshiftConfiguration_blockedDates
=== CONT  TestAccARCZonalShiftZonalAutoshiftConfiguration_blockingAlarms
=== CONT  TestAccARCZonalShiftZonalAutoshiftConfiguration_allowedWindows
=== CONT  TestAccARCZonalShiftZonalAutoshiftConfiguration_disappears
--- PASS: TestAccARCZonalShiftZonalAutoshiftConfiguration_disappears (204.25s)
--- PASS: TestAccARCZonalShiftZonalAutoshiftConfiguration_blockedWindows (216.37s)
--- PASS: TestAccARCZonalShiftZonalAutoshiftConfiguration_allowedWindows (226.23s)
--- PASS: TestAccARCZonalShiftZonalAutoshiftConfiguration_update (229.15s)
--- PASS: TestAccARCZonalShiftZonalAutoshiftConfiguration_blockingAlarms (240.03s)
--- PASS: TestAccARCZonalShiftZonalAutoshiftConfiguration_basic (243.96s)
--- PASS: TestAccARCZonalShiftZonalAutoshiftConfiguration_blockedDates (253.87s)
PASS
ok      github.com/hashicorp/terraform-provider-aws/internal/service/arczonalshift      259.257s
```
